moved array to util.ts
