+++ src/reducers/shelf/filter.ts
@@ -1,7 +1,7 @@
 import {FILTER_ADD, FILTER_REMOVE} from '../../actions/filter';
 import {Action} from '../../actions/index';
 import {DEFAULT_SHELF_UNIT_SPEC, ShelfUnitSpec} from '../../models/shelf/spec';
-import {insertItemToArray, removeItemFromArray} from './spec';
+import {insertItemToArray, removeItemFromArray} from '../util';
 
 export function filterReducer(shelfSpec: Readonly<ShelfUnitSpec> = DEFAULT_SHELF_UNIT_SPEC,
                               action: Action): ShelfUnitSpec {
