+++ src/reducers/shelf/spec.ts
@@ -12,7 +12,6 @@ import {
   SHELF_SPEC_LOAD
 } from '../../actions/shelf';
 
-
 import {AGGREGATE_OPS} from 'vega-lite/build/src/aggregate';
 import {TIMEUNITS} from 'vega-lite/build/src/timeunit';
 import {isWildcardChannelId} from '../../models';
@@ -20,6 +19,7 @@ import {ShelfAnyEncodingDef, ShelfFieldDef, ShelfFunction, ShelfId, ShelfUnitSpe
 import {autoAddFieldQuery} from '../../models/shelf/index';
 import {DEFAULT_SHELF_UNIT_SPEC, fromSpecQuery} from '../../models/shelf/spec';
 import {toSet} from '../../util';
+import {insertItemToArray, modifyItemInArray, removeItemFromArray} from '../util';
 
 
 export function shelfSpecReducer(shelfSpec: Readonly<ShelfUnitSpec> = DEFAULT_SHELF_UNIT_SPEC,
@@ -210,33 +210,4 @@ function removeEncoding(shelf: Readonly<ShelfUnitSpec>, shelfId: ShelfId):
       }
     };
   }
-}
-
-/**
- * Immutable array splice
- */
-export function removeItemFromArray(array: ReadonlyArray<any>, index: number) {
-  return {
-    item: array[index],
-    array: [
-      ...array.slice(0, index),
-      ...array.slice(index + 1)
-    ]
-  };
-}
-
-export function insertItemToArray<T>(array: ReadonlyArray<T>, index: number, item: T) {
-  return [
-    ...array.slice(0, index),
-    item,
-    ...array.slice(index)
-  ];
-}
-
-export function modifyItemInArray<T>(array: ReadonlyArray<T>, index: number, modifier: (t: Readonly<T>) => T) {
-  return [
-    ...array.slice(0, index),
-    modifier(array[index]),
-    ...array.slice(index + 1)
-  ];
 }
